+++ home/noghartt/features/cli/default.nix
@@ -2,6 +2,7 @@
   imports = [
     ./agents
     ./git.nix
+    ./jujutsu.nix
     ./zsh.nix
     ./tools.nix
     ./editor.nix