+++ home/noghartt/features/desktop/applications.nix
@@ -1,0 +1,11 @@
+{ pkgs, ... }:
+{
+  home.packages = with pkgs; [
+    calibre
+    discord
+    slack
+    spotify
+    todoist-electron
+    zotero
+  ];
+}