+++ home/noghartt/features/desktop/default.nix
@@ -1,11 +1,14 @@
 {
   imports = [
     ./hyprland.nix
+    ./noctalia.nix
+    ./rofi.nix
     ./ghostty.nix
     ./firefox.nix
     ./chrome.nix
     ./obsidian.nix
-    ./dunst.nix
     ./flameshot.nix
+    ./applications.nix
+    ./mimeapps.nix
   ];
 }